code ready need to test asap

+++ local/src/c++/system/mech_names.cc
@@ -3,11 +3,11 @@
 Php::Value UniClient::Database::getGroupMembers(std::string group)
     {
         std::vector<std::string> to_return;
-        for(int j=0;j<this->groups.size();j++)
+        for(unsigned long int j=0;j<this->groups.size();j++)
             {
                 if(this->groups[j].name==group)
                     {
-                    for(int i=0;i<this->groups[j].users.size();i++)
+                    for(unsigned long int i=0;i<this->groups[j].users.size();i++)
                         {
                             to_return.push_back(this->groups[j].users[i].username);
                         };
